Refactoring - the structure of the project was changed
