+++ HospitalToday/Common/Models/Person.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace HospitalToday.Domain.Person
+namespace HospitalToday.Common.Models
 {
     abstract class Person
     {
