+++ HospitalToday/DataAccess/Implementation/ReportRepository.cs
@@ -1,11 +1,12 @@
 ﻿using HospitalToday.Domain;
+using HospitalToday.Domain.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace HospitalToday.Repository
+namespace HospitalToday.DataAccess.Implementation
 {
     class ReportRepository : IRepository<Report>
     {
