+++ HospitalToday/Domain/Factory/Implementation/DoctorFactory.cs
@@ -1,10 +1,13 @@
-﻿using System;
+﻿using HospitalToday.Common.Models;
+using HospitalToday.Domain.Factory;
+using HospitalToday.Domain.Models;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace HospitalToday.Domain.Person.Doctor
+namespace HospitalToday.Domaim.Factory.Implementation
 {
     class DoctorFactory : PersonFactory
     {
