+++ HospitalToday/Domain/Factory/PersonFactory.cs
@@ -1,10 +1,11 @@
-﻿using System;
+﻿using HospitalToday.Common.Models;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace HospitalToday.Domain.Person
+namespace HospitalToday.Domain.Factory
 {
     abstract class PersonFactory
     {
