+++ HospitalToday/Domain/Models/Antiseptic.cs
@@ -1,10 +1,11 @@
-﻿using System;
+﻿using HospitalToday.Common.Models;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace HospitalToday.Domain.Medicine.Antiseptic
+namespace HospitalToday.Domain.Models
 {
     class Antiseptic : Medicine
     {
