+++ HospitalToday/Domain/Models/Report.cs
@@ -1,10 +1,11 @@
-﻿using System;
+﻿using HospitalToday.Common.Models;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace HospitalToday.Domain
+namespace HospitalToday.Domain.Models
 {
     class Report
     {
@@ -12,5 +13,6 @@ namespace HospitalToday.Domain
         public DateTime Date { get; set; }
         public int DoctorId { get; set; }
         public int PatientId { get; set; }
+        public List<Medicine> Medicines {get; set;}
     }
 }
